add album artwork to notifications
This essentially just involves connecting MPD's albumart command to the
image-data hint. Unfortunately, MPD's albumart command does not seem to
send artwork embedded in the audio files, instead, it just sends the
contents of cover.png or cover.jpg files, if they exist.

It would be nice to read embedded album artwork in the future...